+++ config.yaml
@@ -1,4 +1,4 @@
-internet_search: False
+internet_search: True
 model: gpt-4o-mini
 
 system:
